+++ src/main/java/com/weather/dao/UserDao.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface UserDao {
     void save(User user);
-    Optional<User> findById(int id);
+    Optional<User> findById(long id);
     Optional<User> findByUsername(String username);
     boolean existByUsername(String username);
 }
